pyproject.toml
[project]
name = "disney-queue-times"
version = "0.1.0"
description = ""
authors = [
    {name = "Justin Rawlinson",email = "justin.rawlinson@pm.me"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "holidays (>=0.73,<0.74)",
    "typer[all] (>=0.16.0,<0.17.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
pytest = "^8.3.5"

